// frontend/src/api/models/TaskRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TaskRequest
 */
export interface TaskRequest {
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    status: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    priority: string;
    /**
     * 
     * @type {Date}
     * @memberof TaskRequest
     */
    dueDate: Date;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    projectId: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    changeDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRequest
     */
    userId?: string;
}

/**
 * Check if a given object implements the TaskRequest interface.
 */
export function instanceOfTaskRequest(value: object): value is TaskRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('priority' in value) || value['priority'] === undefined) return false;
    if (!('dueDate' in value) || value['dueDate'] === undefined) return false;
    if (!('projectId' in value) || value['projectId'] === undefined) return false;
    return true;
}

export function TaskRequestFromJSON(json: any): TaskRequest {
    return TaskRequestFromJSONTyped(json, false);
}

export function TaskRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'],
        'description': json['description'],
        'status': json['status'],
        'priority': json['priority'],
        'dueDate': (new Date(json['dueDate'])),
        'projectId': json['projectId'],
        'changeDescription': json['changeDescription'] == null ? undefined : json['changeDescription'],
        'userId': json['userId'] == null ? undefined : json['userId'],
    };
}

export function TaskRequestToJSON(value?: TaskRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'status': value['status'],
        'priority': value['priority'],
        'dueDate': ((value['dueDate']).toISOString()),
        'projectId': value['projectId'],
        'changeDescription': value['changeDescription'],
        'userId': value['userId'],
    };
}